Configured the selling and cancelling features

+++ apps/cron.py
@@ -1,4 +1,4 @@
-from apps.models.auction import Item
+from apps.models.auction import Item, Bid
 from django.utils import timezone
 
 
@@ -9,5 +9,15 @@ def checkExpiredAuctions():
 
     # Remove all expired auction from database
     for auction in allExpiredAuctions:
-        auction.delete()
+        
+        # Check the bids for all of the expired auctions
+        # Sell the item for the highest bid if the highest bid's price is equal to or more than the item's price
+        allBidsForThisAuction = Bid.objects.filter(item__id=auction.id).order_by('-bidPrice')
+
+        if allBidsForThisAuction.get().bidPrice >= auction.price:
+            auction.status = 'Sold'
+        else:
+            auction.status = 'Cancelled'
+            auction.delete()
+
